Mahmoud Mustafa Oushi/Q&A_App_Streamlit/requirements.txt
streamlit
faiss-cpu
numpy
groq
sentence-transformers
scikit-learn
PyPDF2
pandas
ipywidgets
python-dotenv

